requirements.txt
EbookLib==0.18
Jinja2==3.1.2
beautifulsoup4==4.12.2
openai==0.27.8
requests==2.31.0
python-dotenv==1.0.0
typer==0.7.0